✨ feat: 기수 serializer 보완
